Fix duplicate port error
In the integration I seem to have introduced a port duplication bug
where ports are allocated twice. This change removes the port allocation
from Network.add_link() and lets Link() do the work.

The reason for letting Link() perform the port allocation is that
the dataplane emulation does not use the Network class.