Attempt to implement a setgain CLI command
This seems unnecessarily painful so is probably the
wrong implementation.

The command seems to be able to set the amplifier gain,
but then the signal updates don't seem to work properly.

Modified tests/unit_testing/monitor/single_link_test.py
so it tests the internal API. The test still fails.