Log propagation error in  Monitor.optical_signals() (#79)
It is an error for an optical signal that is present on a node's
output port to have no state for that node.  We should probably
raise an exception, but at the very least we can print an error
message.

This whole section of code still needs to be reorganized and
cleaned up.

Also add a missing space to auditing code message.