Change OpticalSignal.reset() to reset all node input state
Previously we were only resetting the output state. This means that
if after calling turn_on() a signal didn't propagate as far (due to
reaching the maximum number of iterations for loop detection), we
could arrive at an inconsistent state where an amplifier had input
state but no output state.

Also clarified some comments.

OpticalSignal.reset() is a bit confusing because it resets the
default/launch state to its initialized value (why is it even
changed?) and (optionally) preserves the output state at the
originating component. Is there any case when this would be
different from the launch state?

As noted, I'm still not clear on is why we are updating a signal's
default/launch state every time we set it at any location in the
network. This doesn't seem like logical behavior.

I think we need to sort out the state management. A signal's
default state should probably be its launch state, and it should
be adjusted when a transceiver is (re)configured.