Build/CI refinements
makefile

The 'make depend' target is useful for local development when
you haven't installed the package, and it now also installs
the python 'build' package to enable 'make dist' to work.

run-tests.yaml

- apt update is its own build step
- we now install by installing the 'build' package and
  using 'make insstall'
- we now (sort of) test 'make depend' separately

It's tricky to test 'make depend' separately from 'make install'
so for now we just verify that it doesn't fail.