change signal warnings to print() for now (#82)
Raising exceptions in the simulator generally either terminates
it (if it's running standalone) or terminates processing (if
it's running an RPC/REST request.)

Signals arriving at a terminal before it's configured should
not raise an exception. For now, we print a warning message
but this should perhaps change in the future to a debug or
informational message.

Also raise ValueError (instead of Warning) for bad parameters
passed to assoc_rx_to_channel() and get_noise_figure(), as
ValueError is more appropriate for parameter errors.