makefile cleanup + fix make install
Since pip uninstall can silently fail to uninstall,
add --force for now to force a reinstall.

Also add $(PIP) and $(APT) to make it more consistent
and potentially easier to install on distros without apt.

And add wheel to make depend since it seems to be missing
on github actions.