Test installed package in github CI workflow
By default, make test tests the checked out version in the
current directory.

For the github actions workflow, we want to test
the installed package to make sure that it works.

We select this by setting PYTHONPATH to nothing,
by using PYTHONPATH= make test

(Note important space after PYTHONPATH.)

To make this work we change runtests.sh to run everything
with 'sudo PYTHONPATH=$PYTHONPATH'.

For examples/lroadmring.py to work, the SSL key file is now
referred to with a relative path.