Add node.Splitter()
Splitter() is a simple splitter element that currently does
not model coupling loss (although the split percents may
be adjusted to model it as desired.)

Note that Splitter() copies the signal to the split
port if needed, creating a new signal that originates
at the splitter. We do this in case this signal
needs to propagate through one of the elements that
the original signal propagates through.

- add tests/splittertest.py

- also remove dodgy intermediate signal state storage in
  OpticalSignal

- also make assoc_loc_in/out more reliable by requiring at least
  one parameter

- also remove unused include_optical_signal_in_roadm
  and references to it