WIP: changing WSS variable attenuation
The WSS attenuation is changed to avoid the squashing
effect where high power channels are greatly attenuated
to bring them closer to low power channels.

Also a channel_map parameter is introduced; if it is present,
attenuation to reach the target power will only be applied to
channels in the channel map.